update the links and move some more files to common
