+++ sdk/identity/identity-cache-persistence/test/internal/node/clientCertificateCredential.spec.ts
@@ -5,7 +5,7 @@
 
 import * as path from "path";
 import { ClientCertificateCredential, TokenCachePersistenceOptions } from "@azure/identity";
-import { msalNodeTestSetup } from "../../../../identity/test/msalTestUtils";
+import { msalNodeTestSetup, MsalTestCleanup } from "../../../../identity-common/test/msalTestUtils";
 import { env, isPlaybackMode } from "@azure-tools/test-recorder";
 import { ConfidentialClientApplication } from "@azure/msal-node";
 import { MsalNode } from "../../../../identity/src/msal/nodeFlows/msalNodeCommon";
@@ -14,8 +14,6 @@ import assert from "assert";
 import { createPersistence } from "./setup.spec";
 
 const ASSET_PATH = "assets";
-
-export type MsalTestCleanup = () => Promise<void>;
 
 describe("ClientCertificateCredential (internal)", function (this: Mocha.Suite) {
   let cleanup: MsalTestCleanup;
