+++ sdk/identity/identity-common/README.md
@@ -1,0 +1,39 @@
+# Azure Identity Common client library for JavaScript
+
+An internal support library for the various Azure Identity libraries.
+
+This package contains common code that needs to be shared among the other Azure Identity libraries. It is not published to NPM and is not meant for usage by any other consumers.
+
+## Identity client libraries
+
+The following client libraries use this package:
+
+- [@azure/identity](https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/identity/identity/README.md)
+- [@azure/identity-vscode](https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/identity/identity-vscode/README.md)
+- [@azure/identity-cache-persistence](https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/identity/identity-cache-persistence/README.md)
+
+## Getting started
+
+For information on getting started, please see our [Identity client libraries](#identity-client-libraries).
+
+## Key concepts
+
+For information on key concepts, please see our [Identity client libraries](#identity-client-libraries).
+
+## Examples
+
+For examples, please see our [Identity client libraries](#identity-client-libraries).
+
+## Next steps
+
+For information on next steps, please see our [Identity client libraries](#identity-client-libraries).
+
+## Troubleshooting
+
+If you run into issues while using this library, directly or indirectly, please feel free to [file an issue](https://github.com/Azure/azure-sdk-for-js/issues/new).
+
+## Contributing
+
+If you'd like to contribute to this library, please read the [contributing guide](https://github.com/Azure/azure-sdk-for-js/blob/main/CONTRIBUTING.md) to learn more about how to build and test the code.
+
+![Impressions](https://azure-sdk-impressions.azurewebsites.net/api/impressions/azure-sdk-for-js%2Fsdk%2Fidentity%2Fidentity-common%2FREADME.png)
